+++ src/H2DE_functions.cpp
@@ -1,5 +1,5 @@
 #include "H2DE_utils.h"
 
-float lerp(float min, float max, float blend) {
+float H2DE_Lerp(float min, float max, float blend) {
     return min + blend * (max - min);
 }